requirements.txt
pdfplumber==0.8.0
pymongo==4.3.3
argparse==1.4.0
progressbar2==4.0.0
Unidecode==1.3.6